Add further condition options for handlers
Something that is currently missing is the ability to test if the http
request is being called with the correct request data
(body/headers/params).

This commit addresses that by giving the option to add request data to
the handlers. This will allow mocks to only responses for specific data
set.

May be worth adding regex matching (e.g. checking that a header/param
exists but not caring about the value)