Fix matching functionality for mocks
An incorrect assumption was made that the request body from the fetch
was a string. Due to this we try to parse the object before moving on.
However, the request body object type depends on the content-type header
.

This commit addresses the above issue in a bare minimum way. This will
probably not work for some content types, but this will be revisited in
future.