Fix broken tests
A few tests are broken (this is why CI should be added... coming soon!).
These break because they are testing against `https://example.com` which
is a real URL. So when expecting a failure with no matching mock, the
response is still successful.

This commit replaces the url to `http://some.url` which seems to be a
bit better.